Move Collection class.
